refactor(kwin-effect): extract screen helpers to plasmazoneseffect/screens.cpp
Pulls 9 screen-id, virtual-screen, and motion-clock methods out of
plasmazoneseffect.cpp: outputScreenId, getWindowScreenId,
resolveEffectiveScreenId, fetchVirtualScreenConfig,
fetchAllVirtualScreenConfigs, onVirtualScreensChanged, clockForOutput,
onScreenAdded, onScreenRemoved. Drops the main file from 4779 to 4391
lines.
