+++ kwin-effect/CMakeLists.txt
@@ -44,6 +44,7 @@ set(plasmazones_effect_SRCS
     plasmazoneseffect/window_identity.cpp
     plasmazoneseffect/mouse_drag.cpp
     plasmazoneseffect/window_filtering.cpp
+    plasmazoneseffect/screens.cpp
     autotilehandler.cpp
     autotilehandler.h
     autotilehandler/signals.cpp
